twitch_types::color

Struct HexColorRef

pub struct HexColorRef(/* private fields */);

Available on crate feature color only.

A color in hex

Implementations

impl HexColorRef

pub const fn from_str(raw: &str) -> &Self

Transparently reinterprets the string slice as a strongly-typed HexColorRef

pub const fn from_static(raw: &'static str) -> &'static Self

Transparently reinterprets the static string slice as a strongly-typed HexColorRef

pub fn into_owned(self: Box<HexColorRef>) -> HexColor

Converts a Box<HexColorRef> into a HexColor without copying or allocating

pub const fn as_str(&self) -> &str

Provides access to the underlying value as a string slice.

impl HexColorRef

pub fn as_cow(&self) -> Cow<'_, HexColorRef>

Get a Cow<'_, HexColorRef>