twitch_types

Module emote

Available on crate feature emote only.

types for emotes

Structs

BadgeSetId

A Badge set ID

BadgeSetIdRef

A Badge set ID

ChatBadgeId

A channel chat badge ID

ChatBadgeIdRef

A channel chat badge ID

EmoteId

A chat Emote ID

EmoteIdRef

A chat Emote ID

EmoteOccurrence

An emote index as defined by eventsub, similar to IRC emotes twitch tag.

EmoteSetId

An Emote Set ID

EmoteSetIdRef

An Emote Set ID

EmoteUrlBuilder

Builder for emote URLs.

Image

Links to the same image of different sizes

Enums

EmoteAnimationSetting

Formats for an emote.

EmoteScale

Scales available for an emote.

EmoteThemeMode

Background themes available for an emote.

Type Aliases

ResubscriptionEmoteDeprecated

An emote index as defined by eventsub, similar to IRC emotes twitch tag.