twitch_types::emote

Type Alias ResubscriptionEmote

pub type ResubscriptionEmote = EmoteOccurrence;

๐Deprecated since 0.4.8: Use EmoteOccurrence instead

Available on crate feature emote only.

An emote index as defined by eventsub, similar to IRC emotes twitch tag.

Aliased Type

struct ResubscriptionEmote {
    pub begin: i64,
    pub end: i64,
    pub id: EmoteId,
}

Fields
begin: i64

The index of where the Emote starts in the text.

end: i64

The index of where the Emote ends in the text.

id: EmoteId

The emote ID.