twitch_types
Enum Collection

pub enum Collection<'c, T: Deref + 'static>
where
    [T]: ToOwned,
{
    Owned(Cow<'c, [T]>),
    Borrowed(Cow<'c, [&'c T]>),
    Ref(Cow<'c, [&'c T::Target]>),
    OwnedString(Cow<'c, [String]>),
    BorrowedString(Cow<'c, [&'c String]>),
    RefStr(Cow<'c, [&'c str]>),
}

Generic collection of an abstracted item.

This is used to abstract over the different types of collections that can be used, such as Vec<T>, &[T], &[&T], &[&str], etc.

In most cases, you can use the Collection::from method to create a collection.

Examples

use twitch_types::{Collection, UserId, UserIdRef};

// A vector of `UserId`s
let c0: Collection<UserId> = Collection::from(vec![UserId::from("1234"), UserId::from("5678")]);
// A vector of `&str`s
let c1: Collection<UserId> = Collection::from(vec!["1234", "5678"]);
// An array of `&str`s
let c2: Collection<UserId> = Collection::from(&["1234", "5678"]);
// A vector of `UserIdRef`s
let c3: Collection<UserId> = Collection::from(vec![
    UserIdRef::from_static("1234"),
    UserIdRef::from_static("5678"),
]);

assert!([c1, c2, c3].iter().all(|c| *c == c0));

use twitch_types::{Collection, UserId, UserIdRef};
// It's also possible to create a collection from a single item, passing it by reference.
let c0: Collection<UserId> = Collection::from(&"1234");
let id = UserId::from("1234");
let c1: Collection<UserId> = Collection::from(&id);

assert_eq!(c0, c1);

use twitch_types::{Collection, UserId, UserIdRef};
// you can also collect from an iterator
let mut iter = std::iter::from_fn(|| Some(UserId::from("1234"))).take(10);
let c0: Collection<UserId> = iter.collect();
let mut iter = std::iter::from_fn(|| Some(UserIdRef::from_static("1234"))).take(10);
let c1: Collection<UserId> = iter.collect();
let mut iter = std::iter::from_fn(|| Some("1234")).take(10);
let c2: Collection<UserId> = iter.collect();
let mut iter = std::iter::from_fn(|| Some(String::from("1234"))).take(10);
let c3: Collection<UserId> = iter.collect();

assert!([c1, c2, c3].iter().all(|c| *c == c0));

Variants

Owned(Cow<'c, [T]>)

A collection over owned items

Borrowed(Cow<'c, [&'c T]>)

A collection over borrowed items

Ref(Cow<'c, [&'c T::Target]>)

A collection over deref items

OwnedString(Cow<'c, [String]>)

A collection over owned string items

BorrowedString(Cow<'c, [&'c String]>)

A collection over borrowed string items

RefStr(Cow<'c, [&'c str]>)

A collection over &str items

Implementations

impl<'c, T: Deref> Collection<'c, T>
where [T]: ToOwned, for<'t> &'t T::Target: From<&'t str>,

pub const EMPTY: Self = _

An empty collection.

pub fn iter(&self) -> CollectionIter<'_, T>

Returns an iterator over the collection.

Examples

use twitch_types::{Collection, UserId, UserIdRef};

let collection: Collection<UserId> = Collection::from(&["1234", "5678"]);
let mut iter = collection.iter();
assert_eq!(iter.next(), Some(UserIdRef::from_static("1234")));
assert_eq!(iter.next(), Some(UserIdRef::from_static("5678")));
assert_eq!(iter.next(), None);

pub fn into_vec(self) -> Vec<T>
where [T]: ToOwned<Owned = Vec<T>>, for<'d, 'd> T: 'static + Clone + From<&'d <T as Deref>::Target> + From<&'d str>,

Converts the collection into a vector.

Examples

use twitch_types::{Collection, UserId};

let collection = Collection::from(vec!["1", "2", "3"]);
let vector: Vec<UserId> = collection.into_vec();

pub fn chunks( &'c self, chunk_size: usize, ) -> impl Iterator<Item = Collection<'_, T>> + '_

Returns chunks of items, similar to slice::chunks

pub fn len(&self) -> usize

Returns the length of the collection.

pub fn is_empty(&self) -> bool

Returns true if the collection is empty.