Struct twitch_types::stream::CharityDonationIdRef

#[repr(transparent)]
pub struct CharityDonationIdRef(_);

Available on crate feature stream only.

A Charity Donation ID

Implementations

impl CharityDonationIdRef

pub const fn from_str(raw: &str) -> &Self

Transparently reinterprets the string slice as a strongly-typed CharityDonationIdRef

pub const fn from_static(raw: &'static str) -> &'static Self

Transparently reinterprets the static string slice as a strongly-typed CharityDonationIdRef

pub fn into_owned(self: Box<CharityDonationIdRef>) -> CharityDonationId

Converts a Box<CharityDonationIdRef> into a CharityDonationId without copying or allocating

pub const fn as_str(&self) -> &str

Provides access to the underlying value as a string slice.

impl CharityDonationIdRef

pub fn as_cow<'a>(&'a self) -> Cow<'a, CharityDonationIdRef>

Get a Cow<'_, CharityDonationIdRef>