Struct twitch_types::stream::TeamIdRef

#[repr(transparent)]
pub struct TeamIdRef(_);

Available on crate feature stream only.

A Team ID

Implementations

impl TeamIdRef

pub const fn from_str(raw: &str) -> &Self

Transparently reinterprets the string slice as a strongly-typed TeamIdRef

pub const fn from_static(raw: &'static str) -> &'static Self

Transparently reinterprets the static string slice as a strongly-typed TeamIdRef

pub fn into_owned(self: Box<TeamIdRef>) -> TeamId

Converts a Box<TeamIdRef> into a TeamId without copying or allocating

pub const fn as_str(&self) -> &str

Provides access to the underlying value as a string slice.

impl TeamIdRef

pub fn as_cow(&self) -> Cow<'_, TeamIdRef>

Get a Cow<'_, TeamIdRef>