Struct twitch_types::time::Timestamp

#[repr(transparent)]
pub struct Timestamp(_);

Available on crate feature timestamp only.

RFC3339 timestamp

Implementations

impl Timestamp

pub fn new(raw: String) -> Result<Self, TimestampParseError>

Constructs a new Timestamp if it conforms to Timestamp

pub const unsafe fn new_unchecked(raw: String) -> Self

Constructs a new Timestamp without validation

Safety

Consumers of this function must ensure that values conform to Timestamp. Failure to maintain this invariant may lead to undefined behavior.

pub fn from_static(raw: &'static str) -> Self

Constructs a new Timestamp from a static reference if it conforms to Timestamp

Panics

This function will panic if the provided raw string is not valid.

pub fn into_boxed_ref(self) -> Box<TimestampRef>

Converts this Timestamp into a Box<TimestampRef>

This will drop any excess capacity.

pub fn take(self) -> String

Unwraps the underlying String value

impl Timestamp

pub fn now() -> Timestamp

Available on crate feature time only.

Create a timestamp corresponding to current time

pub fn today() -> Timestamp

Available on crate feature time only.

Create a timestamp corresponding to the start of the current day. Timezone will always be UTC.