twitch_types::time
Struct TimestampRef

pub struct TimestampRef(/* private fields */);

Available on crate feature timestamp only.

RFC3339 timestamp

Implementations

impl TimestampRef

pub fn from_str(raw: &str) -> Result<&Self, TimestampParseError>

Transparently reinterprets the string slice as a strongly-typed TimestampRef if it conforms to Timestamp

pub const unsafe fn from_str_unchecked(raw: &str) -> &Self

Transparently reinterprets the string slice as a strongly-typed TimestampRef without validating

pub fn from_static(raw: &'static str) -> &'static Self

Transparently reinterprets the static string slice as a strongly-typed TimestampRef if it conforms to Timestamp

Panics

This function will panic if the provided raw string is not valid.

pub fn into_owned(self: Box<TimestampRef>) -> Timestamp

Converts a Box<TimestampRef> into a Timestamp without copying or allocating

pub const fn as_str(&self) -> &str

Provides access to the underlying value as a string slice.

impl TimestampRef

pub fn as_cow(&self) -> Cow<'_, TimestampRef>

Get a Cow<'_, TimestampRef>

impl TimestampRef

pub fn normalize(&self) -> Result<Cow<'_, TimestampRef>, TimestampParseError>

Normalize the timestamp into UTC time.

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z")?;
assert_eq!(time.normalize()?.as_ref(), &time);
let time2 = Timestamp::try_from("2021-07-01T13:37:00-01:00")?;
assert_ne!(time2.normalize()?.as_ref(), &time2);

pub fn is_before<T: ?Sized>(&self, other: &T) -> bool
where Self: PartialOrd<T>,

Compare another time and return self < other.

Examples

use twitch_types::Timestamp;

let time2021 = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
let time2020 = Timestamp::try_from("2020-07-01T13:37:00Z").unwrap();
assert!(time2020.is_before(&time2021));

pub fn to_day(&self) -> Timestamp

Make a timestamp with the time component set to 00:00:00.

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.to_day().as_str(), "2021-07-01T00:00:00Z");

pub fn year(&self) -> &str

Get the year

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.year(), "2021");

pub fn month(&self) -> &str

Get the month

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.month(), "07");

pub fn day(&self) -> &str

Get the day

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.day(), "01");

pub fn hour(&self) -> &str

Get the hour

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.hour(), "13");

pub fn minute(&self) -> &str

Get the minute

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.minute(), "37");

pub fn second(&self) -> &str

Get the second

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.second(), "00");

pub fn millis(&self) -> Option<&str>

Get the millis

Examples

use twitch_types::Timestamp;

let time = Timestamp::try_from("2021-07-01T13:37:00.123Z").unwrap();
assert_eq!(time.millis(), Some("123"));
let time = Timestamp::try_from("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.millis(), None);

impl TimestampRef

pub fn to_utc(&self) -> OffsetDateTime

Available on crate feature time only.

Construct into a OffsetDateTime time with a guaranteed UTC offset.

Panics

This method assumes the timestamp is a valid rfc3339 timestamp, and panics if not.

pub fn to_fixed_offset(&self) -> OffsetDateTime

Available on crate feature time only.

Construct into a OffsetDateTime time.

Panics

This method assumes the timestamp is a valid rfc3339 timestamp, and panics if not.